% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{sages_data}
\alias{sages_data}
\title{Read SAGES I data from network}
\usage{
sages_data(x, frozen = FALSE, frozen_date = NULL)
}
\arguments{
\item{x}{String Abbreviation of the file to read in.
'560' = The file identifying the n=560 sample
'si' = Subject interview
'hi' = Hospital interview
'mr' = Medical record
'cg' = Caregiver interview}

\item{frozen}{Logical Should a frozen file be used?}

\item{frozen_date}{String "YYYY-MM-DD" If frozen=TRUE, this is the date of the frozen file to use.  If unspecified, it will use
the latest frozen file.  The frozen dates are: "2019-04-23", "2021-05-03". If frozen=FALSE, this is the date of the
current file to use.  If unspecified, it will use the latest current file.  The current dates are: "2022-04-11", "2022-04-20".}
}
\value{
Returns a tibble
}
\description{
This function reads in the SAGES I data that is saved on the network. It
will read in either the frozen or non-frozen data.  If the frozen data is selected
it can either read in the latest freeze or from a particular date.
}
\details{
You must be on the Brown network to use this function.
}
\examples{

\dontrun{
  sages_data("si", frozen = TRUE, frozen_date = "2021-05-03")
}

}
