% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{sagesII_data}
\alias{sagesII_data}
\title{Read SAGES II data from network}
\usage{
sagesII_data(x, frozen = FALSE, frozen_date = NULL)
}
\arguments{
\item{x}{String Abbreviation of the file to read in.
\itemize{
 \item 'si' = Subject interview
 \item 'hi' = Hospital interview
 \item 'mr' = Medical record
 \item 'cg' = Caregiver interview
 }}

\item{frozen}{Logical Should a frozen file be used?}

\item{frozen_date}{String "YYYY-MM-DD" If frozen=TRUE, this is the date of the frozen file to use.  If unspecified, it will use
the latest frozen file.  The frozen dates are: "2022-09-19", "2022-10-05".}
}
\value{
Returns a tibble
}
\description{
This function reads in the SAGES II data that is saved on the network. It
will read in either the frozen or non-frozen data.  If the frozen data is selected
it can either read in the latest freeze or from a particular date.
}
\details{
You must be on the Brown network to use this function.
}
\examples{

\dontrun{
  sagesII_data("si", frozen = TRUE, frozen_date = "2022-10-05")
}

}
