% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{sages_data}
\alias{sages_data}
\title{Read SAGES I data from network}
\usage{
sages_data(x, frozen = FALSE, frozen_date = NULL)
}
\arguments{
\item{x}{String Abbreviation of the file to read in.
\itemize{
  \item '560' = The file identifying the n=560 sample
  \item 'si' = Subject interview
  \item 'hi' = Hospital interview
  \item 'mr' = Medical record
  \item 'cg' = Caregiver interview
}}

\item{frozen}{Logical Should a frozen file be used?}

\item{frozen_date}{String "YYYY-MM-DD" This is the date of the frozen file to use.  If unspecified, it will use
the latest "current" or "frozen" file.  The frozen file dates are: "2019-04-23", "2021-05-03". The current file
dates are: "2022-04-11", "2022-04-20".}
}
\value{
Returns a tibble
}
\description{
This function reads in the SAGES I data that is saved on the network. It
will read in either the frozen or non-frozen data.  If the frozen data is selected
it can either read in the latest freeze or from a particular date.
}
\details{
You must be on the Brown network to use this function.
}
\examples{

\dontrun{
  sages_data("si", frozen = TRUE, frozen_date = "2021-05-03")
}

}
